+++ examples/graph/football.py
@@ -43,5 +43,5 @@ G=parse_gml(gml) # parse gml data
 
 print txt
 # print degree for each team - number of games
-for n,d in G.degree(with_labels=True).items():
+for n,d in G.degree_iter():
     print n,d