+++ networkx/classes/function.py
@@ -47,11 +47,11 @@ def edges_iter(G,nbunch=None):
     """
     return G.edges_iter(nbunch)
 
-def degree(G,nbunch=None,with_labels=False):
+def degree(G,nbunch=None,weighted=False):
     """Return degree of single node or of nbunch of nodes.
     If nbunch is ommitted, then return degrees of *all* nodes.
     """
-    return G.degree(nbunch,with_labels=with_labels)
+    return G.degree(nbunch,weighted=weighted)
 
 def neighbors(G,n):
     """Return a list of nodes connected to node n. """
@@ -120,7 +120,7 @@ def degree_histogram(G):
     Note: the bins are width one, hence len(list) can be large
     (Order(number_of_edges))
     """
-    degseq=G.degree()
+    degseq=G.degree().values()
     dmax=max(degseq)+1
     freq= [ 0 for d in xrange(dmax) ]
     for d in degseq: