+++ networkx/classes/tests/test_graph.py
@@ -244,29 +244,29 @@ class TestGraph:
                           
     def test_degree(self):
         G=self.K3
-        assert_equal(G.degree(),[2,2,2])
-        assert_equal(G.degree(with_labels=True),{0:2,1:2,2:2})
+        assert_equal(G.degree().values(),[2,2,2])
+        assert_equal(G.degree(),{0:2,1:2,2:2})
         assert_equal(G.degree(0),2)
-        assert_equal(G.degree(0,with_labels=True),{0:2})
+        assert_equal(G.degree([0]),{0:2})
         assert_raises((KeyError,networkx.NetworkXError), G.degree,-1)
 
     def test_selfloop_degree(self):
         G=networkx.Graph()
         G.add_edge(1,1)
-        assert_equal(G.degree(),[2])
-        assert_equal(G.degree(with_labels=True),{1:2})
+        assert_equal(G.degree().values(),[2])
+        assert_equal(G.degree(),{1:2})
         assert_equal(G.degree(1),2)
-        assert_equal(G.degree(1,with_labels=True),{1:2})
+        assert_equal(G.degree([1]),{1:2})
 
 
     def test_weighted_degree(self):
         G=networkx.Graph()
         G.add_edge(1,2,weight=2)
         G.add_edge(2,3,weight=3)
-        assert_equal(G.degree(weighted=True),[2,5,3])
-        assert_equal(G.degree(weighted=True,with_labels=True),{1:2,2:5,3:3})
+        assert_equal(G.degree(weighted=True).values(),[2,5,3])
+        assert_equal(G.degree(weighted=True),{1:2,2:5,3:3})
         assert_equal(G.degree(1,weighted=True),2)
-        assert_equal(G.degree(1,weighted=True,with_labels=True),{1:2})
+        assert_equal(G.degree([1],weighted=True),{1:2})