+++ networkx/classes/tests/test_multidigraph.py
@@ -256,10 +256,10 @@ class TestMultiDiGraph(TestMultiGraph):
 
     def test_degree(self):
         G=self.K3
-        assert_equal(G.degree(),[4,4,4])
-        assert_equal(G.degree(with_labels=True),{0:4,1:4,2:4})
+        assert_equal(G.degree().values(),[4,4,4])
+        assert_equal(G.degree(),{0:4,1:4,2:4})
         assert_equal(G.degree(0),4)
-        assert_equal(G.degree(0,with_labels=True),{0:4})
+        assert_equal(G.degree([0]),{0:4})
         assert_raises((KeyError,networkx.NetworkXError), G.degree,-1)
 
     def test_degree_iter(self):
@@ -271,10 +271,10 @@ class TestMultiDiGraph(TestMultiGraph):
 
     def test_in_degree(self):
         G=self.K3
-        assert_equal(G.in_degree(),[2,2,2])
-        assert_equal(G.in_degree(with_labels=True),{0:2,1:2,2:2})
+        assert_equal(G.in_degree().values(),[2,2,2])
+        assert_equal(G.in_degree(),{0:2,1:2,2:2})
         assert_equal(G.in_degree(0),2)
-        assert_equal(G.in_degree(0,with_labels=True),{0:2})
+        assert_equal(G.in_degree([0]),{0:2})
         assert_raises((KeyError,networkx.NetworkXError), G.in_degree,-1)
 
     def test_in_degree_iter(self):
@@ -285,10 +285,10 @@ class TestMultiDiGraph(TestMultiGraph):
 
     def test_out_degree(self):
         G=self.K3
-        assert_equal(G.out_degree(),[2,2,2])
-        assert_equal(G.out_degree(with_labels=True),{0:2,1:2,2:2})
+        assert_equal(G.out_degree().values(),[2,2,2])
+        assert_equal(G.out_degree(),{0:2,1:2,2:2})
         assert_equal(G.out_degree(0),2)
-        assert_equal(G.out_degree(0,with_labels=True),{0:2})
+        assert_equal(G.out_degree([0]),{0:2})
         assert_raises((KeyError,networkx.NetworkXError), G.out_degree,-1)
 
     def test_out_degree_iter(self):